requirements.txt
apache-airflow==1.10.10
awscli==1.18.29
boto3==1.12.29
botocore==1.15.29
requests==2.22.0
simple-salesforce==1.0.0
